requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
captcha==0.7.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-ranged-response==0.2.0
django-recaptcha==4.1.0
django-simple-captcha==0.6.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
pillow==11.3.0
propcache==0.3.2
PyJWT==2.10.1
requests==2.32.4
sqlparse==0.5.3
twilio==9.6.5
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
